requirements.txt
# requirements.txt

# 환경변수(.env) 관리를 위한 라이브러리
python-dotenv==0.21.0

# 디스코드 봇 라이브러리 (py-cord)
py-cord==2.4.0

# (선택) API 연동/HTTP 요청
requests==2.28.2

# (선택) 데이터베이스/ORM
# SQLAlchemy==1.4.46
# aiosqlite==0.19.0